Finish authentication

+++ project/main.py
@@ -22,12 +22,12 @@ def showRestaurants():
   restaurants = db.session.query(Restaurant).order_by(asc(Restaurant.name))
   return render_template('restaurants.html', restaurants = restaurants)
 
-#Create a new restaurant
+#Create a new restaurant & assign that restaurannt to the restaurant owner 
 @main.route('/restaurant/new/', methods=['GET','POST'])
 @login_required
 def newRestaurant():
   if request.method == 'POST':
-      newRestaurant = Restaurant(name = request.form['name'])
+      newRestaurant = Restaurant(name = request.form['name'], owner_id=current_user.id)
       db.session.add(newRestaurant)
       flash('New Restaurant %s Successfully Created' % newRestaurant.name)
       db.session.commit()
@@ -40,9 +40,16 @@ def newRestaurant():
 @login_required
 def editRestaurant(restaurant_id):
   editedRestaurant = db.session.query(Restaurant).filter_by(id = restaurant_id).one()
+  
+  #Check the access if it belongs to the owner of this restarant or it is the admin:
+  if current_user.id != editedRestaurant.owner_id and current_user.role != 'admin':
+    flash('You are not authorised to change the name of this restaurant')
+    return redirect(url_for('main.showRestaurants'))
+    
   if request.method == 'POST':
       if request.form['name']:
         editedRestaurant.name = request.form['name']
+        db.session.commit()
         flash('Restaurant Successfully Edited %s' % editedRestaurant.name)
         return redirect(url_for('main.showRestaurants'))
   else:
@@ -53,6 +60,12 @@ def editRestaurant(restaurant_id):
 @login_required
 def deleteRestaurant(restaurant_id):
   restaurantToDelete = db.session.query(Restaurant).filter_by(id = restaurant_id).one()
+
+  #Check the access if it belongs to the owner of this restarant or it is the admin:
+  if current_user.id != restaurantToDelete.owner_id and current_user.role != 'admin':
+    flash('You are not authorised to delete this restaurant')
+    return redirect(url_for('main.showRestaurants'))
+
   if request.method == 'POST':
     db.session.delete(restaurantToDelete)
     flash('%s Successfully Deleted' % restaurantToDelete.name)
@@ -74,6 +87,12 @@ def showMenu(restaurant_id):
 @login_required
 def newMenuItem(restaurant_id):
   restaurant = db.session.query(Restaurant).filter_by(id = restaurant_id).one()
+
+  #Check the access if it belongs to the owner of this restarant or it is the admin:
+  if current_user.id != restaurant.owner_id and current_user.role != 'admin':
+    flash('You are not authorised to create a new menu item for this restaurant')
+    return redirect(url_for('main.showRestaurants'))
+  
   if request.method == 'POST':
       newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id)
       db.session.add(newItem)
@@ -90,6 +109,12 @@ def editMenuItem(restaurant_id, menu_id):
 
     editedItem = db.session.query(MenuItem).filter_by(id = menu_id).one()
     restaurant = db.session.query(Restaurant).filter_by(id = restaurant_id).one()
+    
+    #Check the access if it belongs to the owner of this restarant or it is the admin:
+    if current_user.id != restaurant.owner_id and current_user.role != 'admin':
+        flash('You are not authorised to edit a menu item for this restaurant')
+        return redirect(url_for('main.showRestaurants'))
+
     if request.method == 'POST':
         if request.form['name']:
             editedItem.name = request.form['name']
@@ -112,6 +137,12 @@ def editMenuItem(restaurant_id, menu_id):
 def deleteMenuItem(restaurant_id,menu_id):
     restaurant = db.session.query(Restaurant).filter_by(id = restaurant_id).one()
     itemToDelete = db.session.query(MenuItem).filter_by(id = menu_id).one() 
+
+    #Check the access if it belongs to the owner of this restarant or it is the admin:
+    if current_user.id != restaurant.owner_id and current_user.role != 'admin':
+        flash('You are not authorised to delete a menu item for this restaurant')
+        return redirect(url_for('main.showRestaurants'))
+
     if request.method == 'POST':
         db.session.delete(itemToDelete)
         db.session.commit()
